Begin to retrieve dataSets with clean arch
