+++ src/compositionRoot.js
@@ -1,3 +1,12 @@
-export async function getCompositionRoot() {
-	return { x: "x" }; //CompositionRoot
+import { DHISRepository } from "./data/repositories/DHISRepository.js";
+import { GetSelectedDataSetsUseCase } from "./domain/usecases/GetSelectedDataSetsUseCase.js";
+
+export function getCompositionRoot() {
+	const dhisRepository = new DHISRepository();
+
+	return {
+		dataSets: {
+			getSelected: new GetSelectedDataSetsUseCase(dhisRepository),
+		},
+	};
 }
