--- src/data/repositories/dhisRepository.js
@@ -1,5 +1,0 @@
-export class DHISRepository {
-	get() {
-		console.log("what");
-	}
-}
